Bugfix for error when delegate requests nil overlay (#391)
* Added non-null guards to OperationList.addBefore & addAfter.

There was already a similar guard on the add() method, but not on the
positional methods. This caused problems when using a delegate script to to
selectively apply an overlay. The documented signal for 'no overlay' is to
return nil.

* Prevent a delegate script create null Overlay instances.

Returning nil is used to signal that an Overlay is not required, so the
operation doesn't need to be added to the OperationList