Exceptions are different when called first time v/s afterwards when Kakadu Libraries are absent
This covers *using the current implementation both type of exceptions
First time: UnsatisfiedLinkError
Once initialization attempt was made: NoClassDefFoundError (VM unloads the JPEG2000KakaduImageReader class?)

There are a few issues with this whole thing as stated in #876, but. a new one of the most obvious ones here is also that the reset does not reset the error message ... and bc really that is what is used *(since we have so many ingored catches) to define if a processor is available or not, at least not for this, but for others that could be not available bc of wrong configs, you can't just reset its failed state....

Still committing, since this matches the way other processors are handled when there is a failure and thus consistent with the current architecture.